feat : created another module for rag multi agent feature and made answer formatting similar across all RAG systems
